product/requirements.txt
absl-py==2.2.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
bidict==0.23.1
binaryornot==0.4.4
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
comfy-cli==1.3.8
comfyui_frontend_package==1.15.13
contourpy==1.3.0
cookiecutter==2.6.0
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
Deprecated==1.2.18
einops==0.8.1
embreex==2.17.7.post6
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastdtw==0.3.4
filelock==3.18.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
ftfy==6.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iopath==0.1.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.2
manifold3d==3.0.1
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matrix-client==0.4.0
mdurl==0.1.2
mediapipe==0.10.21
mixpanel==4.10.1
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
portalocker==3.1.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
pycollada==0.9
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-slugify==8.0.4
python-socketio==5.13.0
PyYAML==6.0.2
questionary==2.1.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rtree==1.4.0
ruff==0.11.6
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
semver==3.0.4
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
spandrel==0.4.1
stringzilla==3.12.4
svg.path==6.3
svglib==1.5.1
sympy==1.13.1
tabulate==0.9.0
termcolor==3.0.1
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchsde==0.2.6
torchvision==0.21.0+cu118
tqdm==4.67.1
trampoline==0.1.2
trimesh==4.6.6
triton==3.2.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
uv==0.6.14
vhacdx==0.0.8.post2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.29.post3
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.19.0
zipp==3.21.0
